+++ src/crunch.rs
@@ -22,23 +22,28 @@ use crate::config::{Config, CONFIG};
 use crate::errors::CrunchError;
 use crate::matrix::Matrix;
 use crate::runtimes::{
-    kusama, polkadot,
+    kusama,
+    polkadot,
     support::{ChainPrefix, ChainTokenSymbol, SupportedRuntime},
-    westend,
+    // westend,
 };
 use async_std::task;
 use log::{debug, error, info, warn};
 use rand::Rng;
-use regex::Regex;
 use serde::Deserialize;
 use std::{convert::TryInto, result::Result, thread, time};
 
 use subxt::{
-    ext::sp_core::{crypto, sr25519, Pair as PairT},
-    storage::StorageKey,
+    backend::{
+        legacy::{rpc_methods::StorageKey, LegacyRpcMethods},
+        rpc::RpcClient,
+    },
+    ext::sp_core::crypto,
     utils::AccountId32,
     OnlineClient, PolkadotConfig,
 };
+
+use subxt_signer::{bip39::Mnemonic, sr25519::Keypair};
 
 pub type ValidatorIndex = Option<usize>;
 pub type ValidatorAmount = u128;
@@ -71,23 +76,39 @@ impl MessageTrait for Message {
     }
 }
 
-pub async fn create_substrate_node_client(
+pub async fn _create_substrate_node_client(
     config: Config,
 ) -> Result<OnlineClient<PolkadotConfig>, subxt::Error> {
     OnlineClient::<PolkadotConfig>::from_url(config.substrate_ws_url).await
 }
 
+pub async fn create_substrate_rpc_client_from_config(
+    config: Config,
+) -> Result<RpcClient, subxt::Error> {
+    RpcClient::from_url(config.substrate_ws_url).await
+}
+
+pub async fn create_substrate_client_from_rpc_client(
+    rpc_client: RpcClient,
+) -> Result<OnlineClient<PolkadotConfig>, subxt::Error> {
+    OnlineClient::<PolkadotConfig>::from_rpc_client(rpc_client).await
+}
+
 pub async fn create_or_await_substrate_node_client(
     config: Config,
-) -> (OnlineClient<PolkadotConfig>, SupportedRuntime) {
+) -> (
+    OnlineClient<PolkadotConfig>,
+    LegacyRpcMethods<PolkadotConfig>,
+    SupportedRuntime,
+) {
     loop {
-        match create_substrate_node_client(config.clone()).await {
-            Ok(client) => {
-                let chain = client.rpc().system_chain().await.unwrap_or_default();
-                let name = client.rpc().system_name().await.unwrap_or_default();
-                let version = client.rpc().system_version().await.unwrap_or_default();
-                let properties =
-                    client.rpc().system_properties().await.unwrap_or_default();
+        match create_substrate_rpc_client_from_config(config.clone()).await {
+            Ok(rpc_client) => {
+                let rpc = LegacyRpcMethods::<PolkadotConfig>::new(rpc_client.clone());
+                let chain = rpc.system_chain().await.unwrap_or_default();
+                let name = rpc.system_name().await.unwrap_or_default();
+                let version = rpc.system_version().await.unwrap_or_default();
+                let properties = rpc.system_properties().await.unwrap_or_default();
 
                 // Display SS58 addresses based on the connected chain
                 let chain_prefix: ChainPrefix =
@@ -117,7 +138,19 @@ pub async fn create_or_await_substrate_node_client(
                     chain, config.substrate_ws_url, name, version
                 );
 
-                break (client, SupportedRuntime::from(chain_token_symbol));
+                match create_substrate_client_from_rpc_client(rpc_client.clone()).await {
+                    Ok(client) => {
+                        break (client, rpc, SupportedRuntime::from(chain_token_symbol));
+                    }
+                    Err(e) => {
+                        error!("{}", e);
+                        info!(
+                            "Awaiting for connection using {}",
+                            config.substrate_ws_url
+                        );
+                        thread::sleep(time::Duration::from_secs(6));
+                    }
+                }
             }
             Err(e) => {
                 error!("{}", e);
@@ -128,24 +161,31 @@ pub async fn create_or_await_substrate_node_client(
     }
 }
 
+// /// Helper function to generate a crypto pair from seed
+// pub fn get_from_seed_DEPRECATED(seed: &str, pass: Option<&str>) -> sr25519::Pair {
+//     // Use regex to remove control characters
+//     let re = Regex::new(r"[\x00-\x1F]").unwrap();
+//     let clean_seed = re.replace_all(&seed.trim(), "");
+//     sr25519::Pair::from_string(&clean_seed, pass)
+//         .expect("constructed from known-good static value; qed")
+// }
+
 /// Helper function to generate a crypto pair from seed
-pub fn get_from_seed(seed: &str, pass: Option<&str>) -> sr25519::Pair {
-    // Use regex to remove control characters
-    let re = Regex::new(r"[\x00-\x1F]").unwrap();
-    let clean_seed = re.replace_all(&seed.trim(), "");
-    sr25519::Pair::from_string(&clean_seed, pass)
-        .expect("constructed from known-good static value; qed")
+pub fn get_signer_from_seed(seed: &str, pass: Option<&str>) -> Keypair {
+    let mnemonic = Mnemonic::parse(seed).unwrap();
+    Keypair::from_phrase(&mnemonic, pass).unwrap()
 }
 
 pub struct Crunch {
     runtime: SupportedRuntime,
     client: OnlineClient<PolkadotConfig>,
+    rpc: LegacyRpcMethods<PolkadotConfig>,
     matrix: Matrix,
 }
 
 impl Crunch {
     async fn new() -> Crunch {
-        let (client, runtime) =
+        let (client, rpc, runtime) =
             create_or_await_substrate_node_client(CONFIG.clone()).await;
 
         // Initialize matrix client
@@ -158,12 +198,17 @@ impl Crunch {
         Crunch {
             runtime,
             client,
+            rpc,
             matrix,
         }
     }
 
     pub fn client(&self) -> &OnlineClient<PolkadotConfig> {
         &self.client
+    }
+
+    pub fn rpc(&self) -> &LegacyRpcMethods<PolkadotConfig> {
+        &self.rpc
     }
 
     /// Returns the matrix configuration
@@ -201,8 +246,8 @@ impl Crunch {
         match self.runtime {
             SupportedRuntime::Polkadot => polkadot::inspect(self).await,
             SupportedRuntime::Kusama => kusama::inspect(self).await,
-            SupportedRuntime::Westend => westend::inspect(self).await,
-            // _ => unreachable!(),
+            // SupportedRuntime::Westend => westend::inspect(self).await,
+            _ => unreachable!(),
         }
     }
 
@@ -210,8 +255,8 @@ impl Crunch {
         match self.runtime {
             SupportedRuntime::Polkadot => polkadot::try_crunch(self).await,
             SupportedRuntime::Kusama => kusama::try_crunch(self).await,
-            SupportedRuntime::Westend => westend::try_crunch(self).await,
-            // _ => unreachable!(),
+            // SupportedRuntime::Westend => westend::try_crunch(self).await,
+            _ => unreachable!(),
         }
     }
 
@@ -223,9 +268,10 @@ impl Crunch {
             SupportedRuntime::Kusama => {
                 kusama::run_and_subscribe_era_paid_events(self).await
             }
-            SupportedRuntime::Westend => {
-                westend::run_and_subscribe_era_paid_events(self).await
-            } // _ => unreachable!(),
+            // SupportedRuntime::Westend => {
+            //     westend::run_and_subscribe_era_paid_events(self).await
+            // }
+            _ => unreachable!(),
         }
     }
 }
@@ -336,7 +382,7 @@ pub async fn try_fetch_onet_data(
     let endpoint = if config.onet_api_url != "" {
         config.onet_api_url
     } else {
-        format!("https://{}-onet-api-beta.turboflakes.io", chain_name)
+        format!("https://{}-onet-api.turboflakes.io", chain_name)
     };
 
     let url = format!(
@@ -374,7 +420,7 @@ pub async fn try_fetch_onet_data(
 }
 
 pub fn get_account_id_from_storage_key(key: StorageKey) -> AccountId32 {
-    let s = &key.0[key.0.len() - 32..];
+    let s = &key[key.len() - 32..];
     let v: [u8; 32] = s.try_into().expect("slice with incorrect length");
     v.into()
 }