+++ src/runtimes/mod.rs
@@ -24,4 +24,4 @@
 pub mod kusama;
 pub mod polkadot;
 pub mod support;
-pub mod westend;
+// pub mod westend;